Working flutter driver tests
