+++ packages/firebase_admob/example/test_driver/firebase_admob_test.dart
@@ -1,0 +1,7 @@
+import 'package:flutter_driver/flutter_driver.dart';
+
+void main() async {
+  final FlutterDriver driver = await FlutterDriver.connect();
+  await driver.requestData(null, timeout: const Duration(minutes: 1));
+  await driver.close();
+}
