+++ packages/firebase_admob/test/firebase_admob_test.dart
@@ -1,0 +1,14 @@
+import 'dart:async';
+
+import 'package:flutter_test/flutter_test.dart';
+import 'package:e2e/e2e.dart';
+
+import 'package:firebase_admob/firebase_admob.dart';
+
+void main() {
+  E2EWidgetsFlutterBinding.ensureInitialized();
+
+  group('$FirebaseAdMob', () {
+
+  });
+}
